neubot/ui.py: /api/results2: New query API
Requested by Antonio.  This new query API should be used by default
in the new web interface of Neubot.  We use /api/results2 ad interim
because /api/results is reserved by the old query API.  But, in the
long term, the old API will be removed and the new API will be avail
at /api/results.  (Or, but not very likely, we will provide the new
API at /api/results/2 and the older at /api/results/1).