neubot/database.py: Make sure we don't pass unicode to lower layers
Note that this function might fail if the default encoder is not
able to encode some charaters.  This avoids a warning in streams.py
and the solution we use here is the same that streams.py tries if
it finds that its input is Unicode.  So, no functional change here.

(However, it is not bullet-proof to convert to "utf-8" because the
encoding might fail if there are strange characters.  This is to
say that we can -and must- do much more in the field of Unicode for
neubot.)