neubot/state.py: Augment state with latest rendezvous's timestamp
Requested by Antonio.

To have this information in the XML document that represents the state
of the neubot daemon simplifies a bit the javascript code that must react
to state changes--at least it simplifies the algorithm we employ in the
upcoming new web user interface.  Indeed, if the timestamp is in the
state, we don't have to calculate the time when the test had started in
the javascript, and that removes some complexity.