www/image: location for images used by the new web user interface
We put in this new directory all the images required by the new web
user interface.  Yes, there are duplicates because the two interfaces
are going to live together for a couple of weeks, but, of course, we
will get rid of the duplicates when we will remove the old web user
interface.