neubot/database.py: Don't use from..import for standard library
This patch is part of the background effort to avoid from..import
for names that belong to the standard library.

As side effect of this patch is the following.  If we want to make
neubot/database.py standalone, we need to know which neubot modules
it imports, and assess the feasibility of replacing them.  With this
patch applied to find the imports from neubot just boils down to
look at the from..imports.