net/streams.py: make sure we print exceptions correctly
The patch cf97744f "net.stream: Add debug statements and propagate
changes" was partially wrong because the exception does not exist
anymore when we handle sorecv / sosend return value.

So, move the call to log.exception in a place where the exception exists
in order to print a meaningful stacktrace.

Noted running my off-tree echo client and server for testing purpose.