connector.py: Simplify message codes
I do not understand why they were implemented this way.  Most likely
that there is a good reason that I cannot understand.  So let's change
the code in a way that seems to me simpler.  Eventually, something
will break and I will learn something and I will revert this commit.