connector.py: Setting to None should be the same as del
Maybe I'm missing something, but to me it seems the same to use del
and to set the object to None.  Maybe the object is destroyed immedia-
tely instead of when the garbage collector runs.  I don't know.  Or,
maybe, they actually are just the same.  So, just set everything to
None for uniformity and then we'll see.