neubot/objgraph.py: Import v1.1b of objgraph
This tool is very valuable for debugging leaks and is a magnitude
better than our current debugging strategy that basically boils
down to enable neubot/debug.py and painfully count the number of
__init__ and __del__ to ensure that they match.

BTW, this does not mean that we are going to zap neubot/debug.py
because this tool might indeed be useful for other purposes.  What
this means is that we can start removing all the painful __del__
with pass that are around and that were there just to hint the
code in neubot/debug.py.

Suggested by Roberto D'Auria and tested while working together
to track an alleged memory leak.