Change Pollable API: pollers.py: Rename closing() to closed()
This mirrors more faithfully what actually happens.  Indeed it is
the poller that closes a socket descriptor and the notification
sent to the related stream is that the connection has been closed
and not that it is closing.

This patch just applies the change to net/pollers.py and the
change to net/streams.py is implemented by a separate patch due
to the fact that also streams.py is being changed this days.