neubot/net: Copy code of test.py into streams.py.
Pass two of a nearly zero effort strategy to get rid of the ugly
file neubot/net/test.py.

This file is here because the code it cointains cannot live into
neubot/net/streams.py.  Indeed the code in test.py depends on the
code in streams.py as well as from the code in connectors.py and
listeners.py.  But, in turn, the code in connectionrs.py and in
listeners.py depends on streams.py.  So, here there is a circular
dependency.

To unbreak that, first pass is to move the relevant code from
connectors.py and listeners.py into streams.py.  And the second
pass is to move code from test.py into streams.py.  Finally,
we get rid of test.py.