Makefile: Allow to make a lite release
We define lite release a release that does not involve creating
updates for APT.  There are two possible use cases for this feature.
The former is when you are not running on a Debian system and so
the tools to generate updates for APT are not available.  And the
latter is when you really don't want to create updates for APT
because the release you are blessing just targets developers and
testers--and you don't want ordinary users (that rely on APT)
to update.