Undo streams.py optimizations: Undo delayed unset readable/writable
Today I have learnt the hard-way what Raymond says in `The Art of
Unix Programing', that one should first carefully measure and then
optimize and that premature optimization is the root of all evil.

This optimization wasn't actually optimizing that much and so I
am reverting it because the reverted code is more simple.