Rewrite Stream in streams.py: Use boolean variables and not flags
In the past I have rewritten streams.py to use flags instead of boolean
variables because there were too much boolean variables.  Now that I have
undone some of the optimizations it's possible to return to the booleans.

Note that -surprise!- the code is faster using booleans rather than
flags and so -umpf- this was another place where I screwed up a bit the
streams.py.