streams.py: Allow to encrypt/decrypt data using ARC4
This is possible because we have modified the protocol to create a
new stream and is required by the upcoming BitTorrent test.

Note that this is implemented so that Neubot can run also when the
module that provides ARC4 (PyCrypto) is not avail.  I doubt that will
be the common case, but I think it's worth trying to keep the set of
must-have dependencies as small as possible.