streams.py: Change relation between Stream and Socket/SSLStream
This change is preparatory for a subsequent change that will modify
the way we create a Stream and will allow to specify configuration
parameters, such as the key for ARC4, or the path to SSL certfile.

Before this patch:

- Socket and SSLStream were derived classes of Stream,

- had to implement the soclose(), sosend(), sorecv() methods,

After this patch:

- the name is changed (Socket/SSLWrapper instead of Socket/SSLStream),

- Socket and SSLWrapper are derived classes of object,

- still have to implement soclose() etc as before,

- the Stream will invoke their soclose() etc methods as needed.

Before this patch the protocol to create a new Stream was like this::

  stream = StreamSSL(sock, poller, ...)

or::

  stream = StreamSocket(sock, poller, ...)

The new protocol, instead, is like this::

  so = SSLWrapper(ssl_sock)
  stream = Stream(poller, so, ...)

or::

  so = SocketWrapper(sock)
  stream = Stream(poller, so, ...)

However, users of the stream are not going to notice this because
they always employ the create_stream() wrapper.