streams.py: Allow to specify configuration parameters
This patch modifies the protocol to create a stream in order to
allow for more flexibility.

The previous path modified the protocol to create a new stream so
that it looks like this::

  so = SSLWrapper(ssl_sock)
  stream = Stream(poller, so, fileno, ...)

or

  so = SocketWrapper(sock)
  stream = Stream(poller, so, fileno, ...)

This patch modifies again the protocol and now it looks like this::

  stream = Stream(poller)
  stream.make_connection(sock)
  stream.configure(conf)

Where sock must be a plain socket and via configure() it is now possible
to specify SSL parameters that cause OpenSSL to be used.

The obvious advantage of this new protocol is that you always invoke
the same three functions and the SSL/non-SSL magic is hidden into configure()
that receives a dictionary with the relevant configuration parameters.

This design, in turn, allows for more flexibility and, indeed, we will
add ARC4 capabilities (required by the upcoming BitTorrent test) in a
few commits.

Note that users of the stream are not going to notice all these changes
because they always employ the create_stream() wrapper.