streams.py: Fork Connector and Listener
This is a preparatory change for an upcoming commit that will
modify Connector and Listener in a way that makes it simpler to
implement other protocols, etc.

I have decided to fork the implementation so that the old code
stays there and we are able to upgrade all the tree incrementally.
Note that consumers of the old code always use connect() and
listen(), so the fact that this patch renames the old Connector
and Listener to OldConnector, OldListener is harmless.