streams.py: More structured Connector and Listener
This patch modifies the forked Connector and Listener so that they
become more structured and provide a skeleton to implement most client
and server classes.  The idea is that a server is a subclass of Listener
and the client is a subclass of Connector.  So, I have added all the
methods to notify the Listener and the Connector of various events.