streams.py: Dispatch started_connecting() on EINPROGRESS only
Dispatch `connection in progress` only when the connection is actually
in progress and don't dispatch it when the connection is already complete
after connect().

This is useful to distinguish the case when the connection completes
immediately (should happen for the local host only) and the case when
it takes a little while to establish.