streams.py: Alternate approach for measuring stats
Now a measurer structure might be attached to a stream using the
stream method configure() and the stream will detach from the
measurer setting the measurer member `dead` to True.

This model provides for just one measurer structure per stream and,
implied, there is a parent measurer structure that groups a certain
number of streams.

The ultimate goal is to monitor the download and upload speed of a
group of connections and to provide per connection stats, such as,
say, "the third connection consumed 33% of the bandwidth."

Note that this commit does not touch the already existing code
employed for accounting, so the consumers of such code (basically
the poller and speedtest.py) should not be affected in any way.