streams.py: Add Measurer and VerboserMeasurer
Measurer is a generic object to perform RTT and bandwidth
measurements.  Note that the latter could be done on a group
of connections and that it's possible to monitor the amount
consumed by each connection.

VerboseMeasurer is an object that builds on Measurer and
prints detailed periodic statistics on the standard ouput,
or at your option, another file-like object.

In order to measure the RTT (actually the time required for
connect() to complete, which is an upper bound), you have to
replace all::

  connector.connect(endpoint

with::

  measurer.connect(connector, endpoint)

In order to monitor the bandwidth consumed by a connection
you need to register the attached stream as follows::

  measurer.register_stream(stream)

In both cases, the measurements will be passed to the measurer,
and will be available to print/calculate statistics.