streams.py: Consider that soclose() might raise
Spotted while auditing the code after the changes to the stream
creation protocol.  When we close the socket it's possible for the
closing function to raise exceptions for various reasons and we
don't want such exceptions to propagate.  It makes completely sense
to just ignore them, because the socket is not going to be used
anymore.