Simplify net module: Move SSL code in streams.py
This patch moves the code to setup an SSL stream from listeners.py and
connectors.py to streams.py.  This simplifies net module because now there
is just one file that pulls from the ssl module (if available).

One notable thing is that the new create_stream() must catch SSLError
when it wrap_socket() and re-raise it as a socket.error because the code
in listeners.py and streams.py that invokes create_stream() could not
catch SSLError exceptions since now just streams.py imports ssl.