streams.py: Allow to send() when we are already sending
This commit fixes one of the grey corners of this stream API, i.e.
the fact that was not defined what to do on send() if we were already
sending.

Two decisions where possibile.  To raise an error or to queue the
incoming data and send it later.  I decided for the latter because
this is what BitTorrent expects (in the handshake there might be
two consecutive writes.)

Note that this commit is slightly wrong as long as it's possible to
specify a different callback for each send().  But I don't address
this issue for two reasons.  First, this is going to change.  Second,
the current code never doest that.