streams.py: More structured relation with protocol and parent
This commit modifies the way the stream works, in order to allow
a more structured relation with the protocol and the parent.

To be clear, here with protocol we mean the object that implements
the syntax of a protocol, such as http.handlers.Handler.  And, with
parent we mean the (server or client) object that owns the stream.

The set of changes is as follows:

- add Stream a parent member that should reference the parent,

- the protocol is now supposed to inherit from Stream and to override
  the implementation of the following (empty) methods:

  connection_made()
    Invoked when the connection is established to start the protocol
    finite state machine.

  connection_lost(exception)
    Invoked when the connection is closed (exception=None) or lost
    due to some exception.

  recv_complete(octets)
    Invoked when an asynchronous receive operation completes.  The
    protocol must invoke start_recv(maxlen) in order to start a new
    asynchronous recv.

  send_complete(count)
    Invoked when an asynchronous send operation completes.  The protocol
    must invoke start_send(octets) in order to start a new async send.
    Note that there could be just one async send active at a time, that
    this limit is not actually enforced, and that I am working to a patch
    that should fix this issue,

- the parent has the option to register self with the stream and, in
  this case, the stream will upcall the parent's connection_lost method
  when it disconnects.

Note that this is implemented in a way that does not disrupt already
existing code, such as neubot.http, because it's still possible to use
the stream with the old API.  Of course, this should be fixed at some
point because to have two parallel API might become fragile.  But,
for now it's great because it allows to incrementally upgrade all the
code.