streams.py: Stream/SSLWrapper: propagate exception to Stream
This is a preparatory change for an upcoming commit that will
make the relation between a stream and a protocol handler much
more structured.

Such upcoming commit will modify stream to pass exceptions to
the protocol handler, and this commit makes that possible.

Note that this change amends slightly the semantic in case of
error.  Before this change the return value in case of error was
ERROR, <don't care>.  After this change the return value is now
the error and the exception that occurred.