streams.py: StreamVerboser: default log messages for stream
This is a simple object that implements the default log messages
a stream should print in reaction to various events.  There should
always be an instance of this object around, therefore you can
always invoke its methods if you like the default messages.  Or
you can ignore this object and roll-out your own log messages.