streams.py: Allow to override default socket buffers size
The optimal buffer size for a connection depends on the bandwidth
and on the round trip time.  A good starting point is to allow users
to change the default buffer size and experiment.

Note that we must set such buffers before connect() and listen()
because the window scaling is negotiated during the three-way hand-
shake.  Therefore it would have not effect to setsockopt() for a
different buffer size once the connection is established.