www: Update with respect to possible neubot.js -> index.js rename
On the one hand, it would be a good idea to rename to index.js for
consistency with neubot.js.  On the other hand index.js seems an odd
name to me.  Update the comment to reflect this fact and postpone
the real decision to the future.

Alessio pointed out that usually people puts everything into a
single javascript so the browser does not need to load stuff when
you switch page.  This make sense and also we avoid the above
neubot.js / index.js issue.