Revert all the changes in the branch pu.dbgen
I tried to revert them using a patch in the feature branch but
something went wrong and so, to keep things simple, I am reverting
this using brute force.

The bottom line is that this branch is not yet ready to go in and
so let's reset the history and retry in the next cycle.