Re-apply some bits that I mistakenly removed from next
During the previous release cycle I changed my usual workflow a bit
because it was broken and it created issues for people tracking neubot
next.  But, apparently, I did some mistake and these bits were left
out of next (indeed you can spot that if you diff next with 0.3.3)
and this patches fixes the issue and re-applies this bits with a little
of brute force.