Revert "More clever buffer usage, kill send_pos"
This patch reverts commit f09ad66... "More clever buffer usage, kill
send_pos" because this commit breaks speedtest.

Indeed, speedtest relies on neubot.http.client to measure the number
of sent bytes as follows::

  class Client(SimpleClient):
      def __init__(self, parent):
          SimpleClient.__init__(self)
          self.sending = SimpleStats()
          self.connecting = SimpleStats()
          self.receiving = SimpleStats()
          self.parent = parent

      ...

      def send_progress(self, data):
          self.sending.account(len(data))

      ...

and the more clever buffer usage cause data to be a subset of the
original data in some cases (namely when you cannot send all in one
send()).

I am not happy of this revert because the code in streams.py is
even more fragile after this revert.  In particular the problem is
the interaction between this code and the code that bufferizes
new pieces of data if we are already sending.

What needs to be done is to migrate neubot.http soon to the new model
to get rid of the low level API in streams.py.  I guess this revert will
be reverted again as soon as we migrate neubot.http.