marshal.py: Rules to marshal and unmarshal simple classes
This rules are a refinement of previous rules in config.py and
experimental rules in tools/dbtool.py.  The main consumer of this
rules is going to be the new config.py but there are many other
modules that can import from this one.

Note that this rules do not allow you to marshal or unmarshal
complex stuff such as dictionaries, lists, or classes.  This
is great because there are already well established Python mods
that do that.  Not least json / simplejson.

BTW, the decision not to depend on simplejson on Python 2.5 leads
to stuff like that.  This module is nice and cute but let's remember
to reconsider simplejson if we find ourselves reinventing too much
the wheel on this topic.  In other words, this module is acceptable
as long as its simple and should be replaced by json otherwise.