streams.py: in main() use the new GenericProtocol
The new generic protocol allows to simplify the code a bit because
there is just one class for all what we want to do.

Note that the default is now changed and the server side sends bytes
while the client side discards it.  This is because it is more common
to test the downstream path rather than the upstream one.  Or, well,
this is what I believe people would be more interested to test.