streams.py: Remove unused protocols, connectors, listeners
We have simplified the code to always use the generic protocol,
connector and listener.  Therefore, we can remove all the classes
that are not used anymore.

Note that the new generic protocol does not implement Echo and this
is something which we might want to add in the future.  However, for
my current tests, I just use Discard and Chargen and so it's ok if we
don't implement Echo now.  Particularly because this allows to reduce
the lenght of an already long file.