streams.py: Make sure -D duration=N prints stats for N seconds
The problem is that the callback that prints stats is scheduled to
run every second but there is no guarantee that it would be scheduled
exactly at a second boundary.  In addition, it is rescheduled every
second.  So, after quite some time, it's possible that such callback
is invoked, say, some tens of milliseconds after the next second
boundary.

The scheduling of the function that stops the clients after N seconds
is much more precise, because it is scheduled just once at the beginning
of the program.

The result is that it's possible that the program does not print the
stats of the last second.

This patch provides a quick, dirty, and possibly wrong fix for the
issue, scheduling the stopping function 1/10 seconds later.  This
should work for short tests (I have tested it for a 16 seconds run)
but might not work if the test is very long.

Since this patch is a bit sloppy, I prefer not to meld it into the
previous one.