streams.py: We don't need to pass count to send_complete()
This is a leftover of the previous API and the rationale was to do
that for simmetry with the receive path.

However, since then, I have realized that the send path and the
receive path are not symmetric.  Indeed, data flows downstream in
the send path and upstream in the recv path (call me stupid, but
it took me a while to realize that.)  So, this change makes sense
with respect to the way the send path actually works.

In addition, this will simplify the send path code.  A recent commit
has modified the stream to make it possible to append bytes to a queue
when there is a send operation in progress.  The current code is
broken because it does not account for such queue when reporting the
caller the number of bytes sent.  So, if we remove the need to pass
the total count upstram, we also remove the need to unbreak the code
with respect to accounting the data added in queue.