streams.py: Fork current Stream implementation
Fork current Stream implementation.  We will keep OldStream as-is
and we will use it in OldConnector and OldListener.  And we will
clean-up class Stream.

The rationale of this changes is that it was starting to be a bit
difficult to maintain backward compatibility within the same class
because one has to account of many subtleties.