streams.py: Make sure client always kicks SSL negotiation off
The client-side of an SSL connection must send the HELLO message to start
the negotiation.  This is done automagically by SSL_read and SSL_write.

When the client first operation is a send(), no problem: the socket is
always writable at the beginning and writable() invokes sosend() that
invokes SSL_write() that negotiates.

The problem is when the client first operation is recv(): in this case
the socket would never become readable because the server side is waiting
for an HELLO.

This patch makes sure that the first recv() of the client invokes
readable() that invokes sorecv() that invokes SSL_read() that starts
the negotiation.

The code seems not to be slower after this patch and this is reasonable
because we have just added a small if() check in the recv fastpath.