neubot: Add simplejson 2.1.3 sources
At the beginning, neubot used JSON.  Then I found that JSON support was
not available in the standard library for Python 2.5.  But I wanted Neubot
to support Python 2.5 and I didn't want Neubot to have external deps. So
I used XML rather than JSON.

This was a mistake.  Indeed I spent quite some time to implement XML mar-
shalling and unmarshalling and this is not working as smoothly as it would
have been with JSON.  Also, the javascript code in the web-interface is
more complex than needed because of XML.

The good news is that bad decisions could be reverted.  Also, there are
often better ways to do things.  If Python 2.5 does not have JSON support
we can just drop simplejson in neubot.

For sure we will replace XML with JSON in the web interface.  But I don't
know whether we will also use JSON for neubot to neubot messages or not.
A sensible strategy to update the web interface might be to start using
JSON for new code and then gradually update other pieces.

Many thanks to Alessio Palmero Aprosio that prodded me to revert the no-JSON
decision a lot.

This patch adds the python part of simplejson.  It also adds the copyright
notice (MIT license) and the changelog to the simplejson directory.

This patch is just the first patch in a branch that enables json support
and falls back to simplejson for Python 2.5.  The branch consists of three
patches:

- [PATCH 1/3] neubot: Add simplejson 2.1.3 sources
- [PATCH 2/3] simplejson: Make sure it's possible to import from neubot
- [PATCH 3/3] compat.py: Import simplejson if json not available