simplejson: Make sure it's possible to import from neubot
Simplejson encoder and decoder use `from simplejson.foo ...` which
does not work because neubot deploys simplejson as a subpackage.

This patch modifies simplejson code to do `from foo ...` for each
foo.  This works because foo is in the simplejson directory.  And is
safe because there is no foo in Python stdlib (I checked.)