marshal.py: Implement JSON_marshal()
This patch adds JSON_marshal(), a function that serializes to JSON
"simple" classes.  A simple class is a class that contains only scalar
integers, strings, and floats.

This patch is the first one in a branch that aims to serialize using
JSON the object returned by /api/config.  The branch contains the following
patches:

1. marshal.py: Implement JSON_marshal()
2. marshal.py: Allow to test either XML or JSON marshalling
3. marshal.py: Zap a redundant `continue`
4. marshal.py: `object` is a reserved word
5. marshal.py: Remember to export JSON_marshal
6. marshal.py: endode -> encode
7. config.py: Return a JSON rather than an XML
8. ui.py: /api/config mimetype: text/xml -> application/json
9. marshal.py: Update to use LOG instead of log

Note that now it's always safe to serialize to JSON: for Python >= 2.6
json is part of the standard library and, for Python 2.5, we fallback to
the instance of simplejson distributed with Neubot.