Move neubot.ico icon from win32 to icons
For quite some time neubot.exe has been a console program and we have
used three wrappers to hide the console for the most common operations,
i.e. start neubot, stop neubot, and show the web user interface.

The code that does the trick is below win32/ -- in particular, the
most relevant file is win32-exec.h.  Not rocket science, baby, just
using the Win32 API in a way that cause the console to be hidden.

Unfortunately, this technique is also employed by some malwares out
there and neubot has been identified as malware.  This means game over,
i.e. we cannot trick Win32 API this way anymore and we will just ship
neubot.exe as a windows program -- it would not be possible anymore
to execute it from cmd.exe but, when you click on it, it would not pop
up a black ugly console.

More in detail, this set of patches will:

- [PATCH 1/6] Move neubot.ico icon from win32 to icons
- [PATCH 2/6] setup.py: Use icons/neubot.ico for neubot.exe
- [PATCH 3/6] setup.py: Make neubot.exe a windows executable
- [PATCH 4/6] neubot.nsi: Always invoke neubot.exe
- [PATCH 5/6] Make-dist.bat: Don't install the content of win32
- [PATCH 6/6] Remove win32 directory and all its content