win32.py: Use NT EventLog handler for logging
Under windows logs were basically sent to /dev/null.  This diff makes
it possible to send them somewhere interesting, namely to the EventLog
Handler of Windows NT.

This is the first patch of a branch that adds support for logging
through NT event logger under Windows.  The branch consists of two
patches:

- [PATCH 1/2] win32.py: Use NT EventLog handler for logging
- [PATCH 2/2] log.py: Delete comment: now there is NT EventLog handler