log.py: Simplify the interactive check a bit
The previous check was stricter but who cares whether stderr is a
tty or not?  The check was overzealous and we can avoid that.

Note that this patch is also helpful in the Windows environment,
when we run as a GUI application.  Indeed, in this case, sys.stderr
does not have a fileno attribute and so this patch prevents Python
to raise an exception.