log.py: clarify where BackgroundLogger comes from
We cannot write it explicitly, e.g.::

  from neubot.unix import BackgroundLogger
  from neubot.win32 import BackgroundLogger

because under windows only win32 defines BackgroundLogger and, in a
similar way, under unix only unix defines this class name.

To clarify things a bit, I have decided to put a comment in front of
the couple of imports, to provide readers a clue.