connector.py: Rebase on the new stream
This patch modifies BitTorrent connector to use the new stream instead
of the old stream.  This means that now we can remove the wrappers at
neubot/bittorrent/handler.py.

Note that now BTConnector is a derived class of the new stream and this
means that it doesn't need a `connection` field anymore.

Also note that the old __init__ has been renamed `initialize` because
now __init__ is the one of the stream.  And note that we have split the
initialization of BTConnector from the moment when the object is noti-
fied that the connection is established, i.e. `connection_made`.  This
is to avoid chicken-and-egg troubles.

Another worth-noting difference is that the modified code must call
the method `start_recv` explicitly to receive the next data chunk from
the peer.  This is the way asynchronous I/O works in neubot, i.e. you
need to re-start recv() each time, and was previously hidden by the
above mentioned wrappers.

While on this patch, also remove from the initializer some variables
that appear to be unused in the current code.  Should these variables
be needed in the future, we can always re-add them back to the code.