main.py: New main for BitTorrent
This main behaves like the old one, i.e. the connection is closed just
after the handshake.  But the difference is that this one employs the
BTHandler object, based on the new stream, and that takes advantage of
the code in neubot/options.py.

Now that we have a new main for the BitTorrent plugin, we just need to
update neubot/main.py in order to pull from this new main.  And then we
can safely remove the old main at neubot/bittorrent/wrappers.py.