connector.py: Implement send_piece()
The original BitTorrent used a scheduler to decide when to send a piece
to the peer.  We implement a more raw approach.  When we want to send we
just put a new piece on the wire.

So, this patch adds a function to send a piece to the peer.

Note that this function invokes _send_message() that concatenates message
length, type, index, begin offset, and the actual block in a byte array, and
then sends such byte array.  I am not sure whether this is efficient or not
as this is actually an extra copy of block but blocks are not so big.