bittorrent: mv connector.py streams.py for simmetry with neubot.net
Rename the py file that contains the BitTorrent stream streams.py,
for simmetry with what happens for neubot/net.  BTW, probably it would
have been better to name the original file stream.py and not
streams.py, but, after all, this makes little difference, and I
believe we can also live with the plural.