state.py: Rewrite to reduce complexity and use JSON
This patch rewrites state.py to remove most of the complexity and,
while on that, to serialize the state using JSON.

State.py was trying to enforce a meaning to the data passed by other
Neubot modules.  This design violates the end to end principle and
unsurprisingly the resulting code was complex to read and it was difficult
to make changes.  The first developer to notice that was Roberto D'Auria,
when he wrote a patch to pass the `rendezvous failed' event to the
web user interface.  He prodded me to improve the code in state.py,
but eventually nothing changed until today, due to other priorities.
The need to serialize the state using JSON prompted the issue of state.py
complexity again.  And this time I've had no mercy of the existing code
and I am rewriting it with this patch.

The new state is a dictionary like this::

  {
    "event_name": event_value
  }

Where each event is versioned -- the version being the opaque time of
the last occurrence of such event.  And where the code returns just the
events older than a given opaque time, if specified.