notify.py: Use all-uppercase for global objects
This is a new style rule I'm trying to push.  The rationale is
that it's crystal clear than something is global when its name is
all-uppercase.

This patch is the first one in a three-patches branch that aims
to enhance <notify.py>, formalizing the concept of `opaque timer`.
That is the timer used to clock internal events, such as state
change, and that, ultimately, drives the Comet.  We need to add
this new opaque time concept because that would help to implement
a simpler version of <state.py>.  More in detail the branch is
composed of the following patches:

1. notify.py: Use all-uppercase for global objects
2. notify.py: Allow to pass publish() the opaque time