api_client.py: Add client that tracks /api/state using comet
This is a rewrite of SimpleStateTracker in <neubot/ui.py> that
expects the state to be JSON rather than XML.

The reason why this code is deployed in a new file instead of
neubot/ui.py is that ui.py contains too much code with different
purposes.  So, I think it was cleaner to deploy the new code in
a separate file, to reduce <ui.py>'s mess.