marshal.py: Implement XML_unmarshal()
This patch implements XML_unmarshal(), a function that unmarshals
XML documents to "simple" classes.  A simple class is a class that
contains only scalar integers, strings, and floats.

This patch is the first one in a branch that aims to serialize the
results of speedtest using JSON rather than XML.  The branch contains
the following patches:

1. marshal.py: Implement XML_unmarshal()
2. database.py: Allow to export speedtest results using JSON
3. ui.py: Add /api/speedtest to pull JSON results
4. ui.py: Bye bye /api/results /api/results2
5. database.py: Bye bye cache and old methods to access results

We need XML_unmarshal() in tree because we need to convert results
in the database from XML to JSON.  And this function provides a very
convenient way to perform part of the conversion.