database.py: Allow to export speedtest results using JSON
This patch is very unclean because the results are stored in the
database as XML documents.  It would become more clean when results
will be stored as plain columns of an SQL database.  (And, yes,
it would have been better to use plain columns in the first place,
and, yes, it was a bad design decision not to do that.)