bittorrent/main.py: Simplest possible transmission test
This patch implements in bittorrent/main.py the simplest possible
transmission test.  The two peers exchange pieces consisting of 32768
times the character 'A'.

This patch is the first one in a branch that aims to enhance the code
in neubot/bittorrent to implement a workable transmission test.  The
branch contains the following patches:

1. bittorrent/main.py: Simplest possible transmission test
2. bittorrent/main.py: Use arcfour_new to scramble pieces
3. bittorrent/main.py: Make piece scrambling conditional
4. bittorrent/main.py: Allow to obfuscate the whole protocol
5. bittorrent/streams.py: Pass CHOKE/UNCKOKE INTERESTED/NOT_INTERESTED
6. bittorrent/main.py: Honour CHOKE/UNCHOKE INTERESTED/NOT_INTERESTED
7. bittorrent/main.py: Keep at most 32 pieces in flight
8. bittorrent/main.py: Allow to ovverride default sobufs size
9. bittorrent/streams.py: Increase robustness and verbosity

The simple test we implement in this patch is clearly not enough and
subsequent patches try to refine this model a bit.  In particular, it
would be great the keep more than one message at a time in flight in
between the two peers.