arcfour.py: New home for ARC4 code
This patch copies the (small amount of) ARC4 code that lives into
net/streams.py in arcfour.py and adds a simple unit test to measure
the throughput.

This patch is the first one in a branch that moves ARC4 code into a
separate file and ensures that such code works.  The branch contains
the following patches:

1. arcfour.py: New home for ARC4 code,
2. net/streams.py: Import ARC4 code from arcfour.py
3. net/streams.py: Unbreak ARC4

We move the ARC4 code in a shared place because the BitTorrent module
is going to be amended to use such code.

A couple of notes:

- if you don't provide a key the word 'neubot' is used, so the obfuscated
  stream will always be the same.  This might be good if you are testing, but
  is definitely bad if you want to hide,

- in the unit test, the measured throughput depends on the size of the
  message that you pass to encrypt().  For example, on my desktop, I get
  1.1 Gbit/s with 32768, 716 Mbit/s with 32, and 879 Mbit/s with 32 M,

- py-Crypto is still an optional dependency, therefore we implement a
  fallback for when it is not available.  There are plans for py-Crypto,
  but, as of now, it's not crystal clear whether this Python extension
  will become mandatory for the client-side or not.